Move `HTTPClientStub` & `InMemoryFeedStore` into a separate files in helpers folder
