+++ EssentialApp/EssentialAppTests/FeedAcceptanceTests.swift
@@ -77,74 +77,6 @@ class FeedAcceptanceTests: XCTestCase {
         sut.sceneWillResignActive(UIApplication.shared.connectedScenes.first!)
     }
     
-    private class InMemoryFeedStore: FeedStore, FeedImageDataStore {
-        private(set) var feedCache: CachedFeed?
-        private var feedImageDataCache = [URL: Data]()
-        
-        static var empty: InMemoryFeedStore {
-            InMemoryFeedStore()
-        }
-        
-        static var withExpiredCache: InMemoryFeedStore {
-            let store = InMemoryFeedStore()
-            store.feedCache = ([], Date.distantPast)
-            return store
-        }
-        
-        static var withNonExpiredCache: InMemoryFeedStore {
-            let store = InMemoryFeedStore()
-            store.feedCache = ([], Date())
-            return store
-        }
-        
-        func retrieve(completion: @escaping FeedStore.RetrievalCompletion) {
-            completion(.success(feedCache))
-        }
-        
-        func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping FeedStore.InsertionCompletion) {
-            feedCache = (feed, timestamp)
-        }
-        
-        func deleteCachedFeed(completion: @escaping FeedStore.DeletionCompletion) {
-            feedCache = nil
-            completion(.success(()))
-        }
-        
-        func retrieve(dataForURL url: URL, completion: @escaping (FeedImageDataStore.RetrievalResult) -> Void) {
-            completion(.success(feedImageDataCache[url]))
-        }
-        
-        func insert(data: Data, forURL url: URL, completion: @escaping (FeedImageDataStore.InsertionResult) -> Void) {
-            feedImageDataCache[url] = data
-            completion(.success(()))
-        }
-    }
-    
-    private class HTTPClientStub: HTTPClient {
-        private let stub: (URL) -> HTTPClient.Result
-        
-        private class Task: HTTPClientTask {
-            func cancel() {}
-        }
-        
-        init(stub: @escaping (URL) -> HTTPClient.Result) {
-            self.stub = stub
-        }
-        
-        func get(from url: URL, completion: @escaping (HTTPClient.Result) -> ()) -> EssentialFeed.HTTPClientTask {
-            completion(stub(url))
-            return Task()
-        }
-        
-        static func online(_ stub: @escaping (URL) -> (HTTPURLResponse, Data)) -> HTTPClientStub {
-            HTTPClientStub { url in .success(stub(url)) }
-        }
-        
-        static var offline: HTTPClientStub {
-            HTTPClientStub { _ in .failure(NSError(domain: "offline", code: 0)) }
-        }
-    }
-    
     private func response(for url: URL) -> (HTTPURLResponse, Data) {
         let response = HTTPURLResponse(url: url, statusCode: 200, httpVersion: nil, headerFields: [:])!
         return (response, makeData(for: url))
