+++ EssentialApp/EssentialAppTests/Helpers/InMemoryFeedStore.swift
@@ -1,0 +1,52 @@
+//
+//  InMemoryFeedStore.swift
+//  EssentialAppTests
+//
+//  Created by Andrey on 9/23/22.
+//
+
+import Foundation
+import EssentialFeed
+
+class InMemoryFeedStore: FeedStore, FeedImageDataStore {
+    private(set) var feedCache: CachedFeed?
+    private var feedImageDataCache = [URL: Data]()
+    
+    static var empty: InMemoryFeedStore {
+        InMemoryFeedStore()
+    }
+    
+    static var withExpiredCache: InMemoryFeedStore {
+        let store = InMemoryFeedStore()
+        store.feedCache = ([], Date.distantPast)
+        return store
+    }
+    
+    static var withNonExpiredCache: InMemoryFeedStore {
+        let store = InMemoryFeedStore()
+        store.feedCache = ([], Date())
+        return store
+    }
+    
+    func retrieve(completion: @escaping FeedStore.RetrievalCompletion) {
+        completion(.success(feedCache))
+    }
+    
+    func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping FeedStore.InsertionCompletion) {
+        feedCache = (feed, timestamp)
+    }
+    
+    func deleteCachedFeed(completion: @escaping FeedStore.DeletionCompletion) {
+        feedCache = nil
+        completion(.success(()))
+    }
+    
+    func retrieve(dataForURL url: URL, completion: @escaping (FeedImageDataStore.RetrievalResult) -> Void) {
+        completion(.success(feedImageDataCache[url]))
+    }
+    
+    func insert(data: Data, forURL url: URL, completion: @escaping (FeedImageDataStore.InsertionResult) -> Void) {
+        feedImageDataCache[url] = data
+        completion(.success(()))
+    }
+}
